Organise growing amount of mjs files in dirs
