+++ src/client/resources/web/js/managers/favicon_manager.mjs
